hotfix for question answer evaluation

+++ sources/database/database_structures.cpp
@@ -2,8 +2,9 @@
 
 std::string data::AnswerBase::tableName         = {"answer"};
 std::vector<data::Type> data::AnswerBase::types = {
-    data::Type::INT,    data::Type::INT, data::Type::INT,   data::Type::STRING,
-    data::Type::STRING, data::Type::INT, data::Type::STRING};
+    data::Type::INT,    data::Type::INT,    data::Type::INT,
+    data::Type::STRING, data::Type::STRING, data::Type::FLOAT,
+    data::Type::STRING};
 std::vector<std::string> data::AnswerBase::columnNames = {
     "id", "user_id", "question_id", "time", "verdict", "weight", "value"};
 std::unordered_map<std::string, uint8_t> data::AnswerBase::nameToNum = {
@@ -407,7 +408,7 @@ data::ProblemBase::reset()
 std::string data::QuestionBase::tableName         = {"question"};
 std::vector<data::Type> data::QuestionBase::types = {
     data::Type::INT,    data::Type::STRING, data::Type::STRING,
-    data::Type::STRING, data::Type::INT,    data::Type::STRING};
+    data::Type::STRING, data::Type::FLOAT,  data::Type::STRING};
 std::vector<std::string> data::QuestionBase::columnNames = {
     "id", "name", "nickname", "type", "weight", "jury_answer"};
 std::unordered_map<std::string, uint8_t> data::QuestionBase::nameToNum = {
